Initial Commit of 'worker node'
Simple vagrant worker node which launches the main.go worker in a docker
container and then runs it. Never used chef before, so probably got some
resource notification fixes to do and ugliness to fix in the recipe -
but it's a start.

Listens on :8484 inside the container but the firewall is still enabled
at this point - as is the default in CentOS 7.