Fix up firewall in deployment
Allow the firewall to be configured. Not all cloud OS images have the
firewall disabled by default, so let's just assume it is there and
configure it.

The nginx module in the global Berksfile was superfluous so that has
also been removed. If we want it again we can use a metadata.rb within
the relevant cookbook.