Stub balancer role + multi-machine Vagrantfile
Update the Vagrantfile to support multiple machines with a 'count' style
config so that someone is able to modify the count of VMs to make.
Obviously the way this is written at the moment is very tied to Vagrant
with a Virtualbox style setup where the IPs can be configured - however
that serves as a good test bed.