Complete nginx load balancer set on top of workers
Does what says on the tin. If you `vagrant up` you'll have a worker-1
and worker-2 with a load balancer in front of them. Obviously the
hostname returned by the Go program is that of the container. We can fix
that later with an environment variable to pass in.

Refactored the Vagrantfile a smidge to pass in the IP addresses of the
workers. That appears to work quite well - just don't randomly re-order
the Hash containing the hosts.

Still takes ages to build all this stuff because ultimately we don't use
Packer to prepare the workers. That's something which would make this
quicker.